# openapi.yaml
openapi: 3.1.0
info:
  title: saas-scaffolding
  version: 1.0.0
  description: Multi-tenant SaaS scaffolding API with authentication, organization management, microservice tokens, and webhook integrations.
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie set after successful sign-in.
  schemas:
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Internal server error
    ErrorResponseWithError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Unauthorized
...
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: jane@example.com
    ResetPasswordRequest:
      type: object
      required:
        - token
        - password
      properties:
        token:
          type: string
          format: uuid
        password:
          type: string
          format: password
          minLength: 8
          description: Must contain at least one uppercase letter, one lowercase letter, one number, and one special character.
    VerifyEmailRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          format: uuid
    Verify2FARequest:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 6
          description: 6-digit verification code.
          example: "123456"
...
    ResendVerificationRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
...
    UserOrganization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        plan:
          type:
            - string
            - "null"
        logo_url:
          type:
            - string
            - "null"
        role:
          type: string
          enum:
            - owner
            - admin
            - member
    GetUserOrgsResponse:
      type: object
      required:
        - organizations
      properties:
        organizations:
          type: array
          items:
            $ref: "#/components/schemas/UserOrganization"
...
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    RefreshedTokenResponse:
      type: object
      required:
        - accessToken
        - expiresIn
        - tokenType
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
          example: 3600
        tokenType:
          type: string
          example: Bearer
...
  responses:
    BadRequest:
      description: Bad request - validation error or missing required fields.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or session invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /api/auth/forgot-password:
    post:
      operationId: forgotPassword
      summary: Request a password reset link
      description: |
        Generates a password reset token and sends an email. Always returns 200 to avoid leaking whether the email exists.
        Rate limit tier: strictAuthRateLimit.
      tags:
        - auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Reset link sent (or no-op if email not found).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: If an account with that email exists, we have sent a password reset link.
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # Auth - Reset Password
  # ──────────────────────────────────────────────
  /api/auth/reset-password:
    post:
      operationId: resetPassword
      summary: Reset password with token
      description: |
        Resets the user password using a previously issued reset token. Validates password strength.
        Rate limit tier: strictAuthRateLimit.
      tags:
        - auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: Password reset successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # Auth - Verify Email
  # ──────────────────────────────────────────────
  /api/auth/verify-email:
    post:
      operationId: verifyEmail
      summary: Verify email address
      description: |
        Confirms a user's email using the verification token sent during signup.
        Rate limit tier: strictAuthRateLimit.
      tags:
        - auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequest"
      responses:
        "200":
          description: Email verified successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: Email verified successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # Auth - Verify 2FA (session-based)
  # ──────────────────────────────────────────────
  /api/auth/verify-2fa:
    post:
      operationId: verify2FA
      summary: Verify a 2FA code (authenticated)
      description: |
        Verifies a 6-digit 2FA code for the currently authenticated user session.
        Rate limit tier: authRateLimit.
      tags:
        - auth
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Verify2FARequest"
      responses:
        "200":
          description: 2FA verification successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: 2FA verification successful
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # Auth - Complete 2FA (login flow)
  # ──────────────────────────────────────────────
...
  /api/auth/resend-verification:
    post:
      operationId: resendVerification
      summary: Resend email verification link
      description: |
        Generates a new verification token and re-sends the verification email. Fails if the email is already verified.
        Rate limit tier: strictAuthRateLimit.
      tags:
        - auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendVerificationRequest"
      responses:
        "200":
          description: Verification email sent successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: Verification email sent successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # Auth - Resend 2FA Code
  # ──────────────────────────────────────────────
...
  /api/auth/get-user-orgs:
    get:
      operationId: getUserOrgs
      summary: List organizations for the current user
      description: |
        Returns all active organizations the authenticated user belongs to, including role and logo.
        Also triggers auto-accept domain membership checks.
        Rate limit tier: apiRateLimit.
      tags:
        - auth
      security:
        - sessionCookie: []
      responses:
        "200":
          description: List of user organizations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetUserOrgsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # Auth - Providers
  # ──────────────────────────────────────────────
...
  /api/microservices/refresh:
    post:
      operationId: refreshMicroserviceToken
      summary: Refresh a microservice access token
      description: |
        Exchanges a valid refresh token for a new JWT access token (1 hour).
        The user must still have an active session and the refresh token must match the session user.
        Rate limit tier: apiRateLimit.
      tags:
        - microservices
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: New access token issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshedTokenResponse"
        "400":
          description: Missing refresh token or no organization context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseWithError"
        "401":
          description: Invalid refresh token or session mismatch.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseWithError"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  # ──────────────────────────────────────────────
  # Microservices - Revoke Token
  # ──────────────────────────────────────────────
...