# src/app/api/auth/verify-email/route.ts
import { NextRequest, NextResponse } from 'next/server';

import { verifyEmailToken } from '@/lib/auth';
import { logger } from '@/lib/debug';
import { strictAuthRateLimit } from '@/lib/rate-limit';

export async function POST(request: NextRequest) {
  const limited = strictAuthRateLimit(request);
  if (limited) return limited;

  try {
    const { token } = await request.json();

    if (!token) {
      return NextResponse.json(
        { message: 'Verification token is required' },
        { status: 400 }
      );
    }

    const result = await verifyEmailToken(token);

    if (!result.success) {
      return NextResponse.json(
        { message: result.message },
        { status: 400 }
      );
    }

    return NextResponse.json(
      { message: 'Email verified successfully' },
      { status: 200 }
    );
  } catch (error: unknown) {
    logger.error('Email verification error:', { error: error instanceof Error ? error.message : error });
    
    return NextResponse.json(
      { message: 'Internal server error' },
      { status: 500 }
    );
  }
}